Add files via upload
Wrote the current bugs up top, most are spawning from sensor errors, so AJ and Ryan are going to work on implementing new better sensors.

The code itself seems extremely consistent when the sensors are reading right.  Might need some more optimization for reading the turn.  That would be something like recording the time of each leg and slowing down the robot a little before when the turn is going to be.